wip: implement GET employees route handler with axios
