+++ src/app/api/employees/route.ts
@@ -1,5 +1,9 @@
 import { getToken } from 'next-auth/jwt'
 import { NextRequest } from 'next/server';
+import { backend } from '@/lib/services/backendService';
+import { edgeBackend } from '@/lib/services/edgeBackend';
+import { Employee } from '@/lib/api/employees.d';
+import { AxiosError, AxiosResponse } from 'axios';
 
 const baseURL = process.env.NEXT_PUBLIC_BACKEND_URL;
 
@@ -12,30 +16,14 @@ export async function GET(req: NextRequest) {
   }
 
   try {
-    const response = await fetch(`${baseURL}/api/employees`,{
-      headers: { Authorization: `Bearer ${jwtToken.accessToken}` },
-    });
-
-    if(!response.ok) {
-      // const errorData = await response.json();
-
-      if (response.status === 400) {
-        return createResponse(400, { message: 'Bad Request' });
-      } else if (response.status === 404) {
-        return createResponse(404, { message: 'Resource Not Found' });
-      } else {
-        // Generic error handler for other statuses
-        return createResponse(response.status, { message: 'An error occurred' });
-      }
-    }
-
-    const data = await response.json();
-    return createResponse(response.status, data);
+    const res: AxiosResponse = await edgeBackend.get(`/api/employees`, { accessToken: 'jwtToken.accessToken' as string, fullResponse: true });
+    return createResponse(res.status, res.data);
 
   } catch (error: any) {
-    // Handle network errors or unexpected errors
-    console.error('Error fetching data from backend:', error);
-    return createResponse(500, { message: 'Internal Server Error', details: error.message });
+    const status = error.response?.status;
+    const message = error.response?.data?.message;
+    const details = error.response?.data;
+    return createResponse(status || 500, { message, details });
   }
 }
